2019/10/02 : Add GCC support.

+++ BSP/Driver/Source/SYS/wb_uart.c
@@ -46,8 +46,8 @@ UINT8 uart_rx[RX_ARRAY_NUM] = {0};
 VOID _PutChar_f(UINT8 ucCh);
 UINT32 volatile rx_cnt = 0;
 
-#define sysTxBufReadNextOne()	(((_sys_uUartTxHead+1)==UART_BUFFSIZE)? NULL: _sys_uUartTxHead+1)
-#define sysTxBufWriteNextOne()	(((_sys_uUartTxTail+1)==UART_BUFFSIZE)? NULL: _sys_uUartTxTail+1)
+#define sysTxBufReadNextOne()	(((_sys_uUartTxHead+1)==UART_BUFFSIZE)? (UINT32)NULL: _sys_uUartTxHead+1)
+#define sysTxBufWriteNextOne()	(((_sys_uUartTxTail+1)==UART_BUFFSIZE)? (UINT32)NULL: _sys_uUartTxTail+1)
 #define UART_BUFFSIZE	256
 UINT8 _sys_ucUartTxBuf[UART_BUFFSIZE];
 UINT32 volatile _sys_uUartTxHead, _sys_uUartTxTail;
@@ -326,13 +326,13 @@ INT32 sysInitializeUART(WB_UART_T *uart)
 	// hook UART interrupt service routine
 	if (u32UartPort)
 	{//==1 NORMAL UART
-		_sys_uUartTxHead = _sys_uUartTxTail = NULL;
+		_sys_uUartTxHead = _sys_uUartTxTail = (UINT32)NULL;
 		_sys_pvOldUartVect = sysInstallISR(IRQ_LEVEL_1, IRQ_UART, (PVOID)sysUartISR);
 		sysEnableInterrupt(IRQ_UART);		
 	}
 	else
 	{//==0 High SPEED
-		_sys_uUartTxHead = _sys_uUartTxTail = NULL;
+		_sys_uUartTxHead = _sys_uUartTxTail = (UINT32)NULL;
 		_sys_pvOldUartVect = sysInstallISR(IRQ_LEVEL_1, IRQ_HUART, (PVOID)sysUartISR);
 		sysEnableInterrupt(IRQ_HUART);
 	}
